Expiry: LockTime component for reward form
