+++ src/3_features/feed/createRewardForm/ui/CreateRewardForm.tsx
@@ -11,26 +11,9 @@ import { useRouter } from 'next/navigation'
 import s from './CreateRewardForm.module.css'
 import { hintsConfig } from '@/5_shared/config/hints.config'
 import { catchHTTPValidationError } from '@/5_shared/utils/catchHTTPValidationError'
+import { LockTimeSelector } from '@/5_shared/ui/LockTimeSelector'
+import { convertToLockedUntilMins } from '@/5_shared/utils/timeConversion'
 
-const timeUnits = [
-    { label: 'Minutes', value: 'minutes' },
-    { label: 'Hours', value: 'hours' },
-    { label: 'Days', value: 'days' },
-    { label: 'Weeks', value: 'weeks' },
-    { label: 'Months', value: 'months' },
-]
-
-const convertToLockedUntilMins = (amount: number, unit: string): number => {
-    const multipliers: Record<string, number> = {
-        minutes: 1,
-        hours: 60,
-        days: 1440,
-        weeks: 10080,
-        months: 43200 // Approximation of a 30-day month
-    }
-
-    return amount * (multipliers[unit] ?? 1)
-}
 
 const CreateRewardForm: FC = () => {
     const queryClient = useQueryClient();
@@ -151,48 +134,16 @@ const CreateRewardForm: FC = () => {
                                 </Col>
                             </Row>
 
-                        <Collapse
-                            bordered={false}
-                            style={{ background: 'none', padding: 0, marginTop: '8px' }} // optional styling
-                        >
-                        <Panel key="1"  header="Advanced settings">
-                            <Row gutter={[8, 8]} align="middle" style={{ marginTop: '8px' }}>
-
-                                <Col span={8}  style={{ display: 'flex', alignItems: 'center' }} >
-                                    <Typography.Text>
-                                        Lock reward until:
-                                    </Typography.Text>
-                                </Col>
-
-                                <Col span={4}>
-                                    <Form.Item
-                                        name="lockedUntilAmount"
-                                        rules={[
-                                            { required: true, message: 'Required field' },
-                                            { type: 'number', min: 1, message: 'Value must be at least 1' }
-                                        ]}
-                                        normalize={(value: string) => Number(value)}
-                                    >
-                                        <Input type="number" min={1} placeholder="Duration" />
-                                    </Form.Item>
-                                </Col>
-                                <Col span={6}>
-                                    <Form.Item
-                                        name="lockedUntilUnit"
-                                        rules={[
-                                            { required: true, message: 'Required field' }
-                                        ]}
-                                    >
-                                        <Select
-                                            options={timeUnits}
-                                            placeholder="Select unit (e.g. weeks)"
-                                        />
-                                    </Form.Item>
-                                </Col>
-                                </Row>
+                            <Collapse
+                                bordered={false}
+                                style={{ background: 'none', padding: 0, marginTop: '8px' }} // optional styling
+                                >
+                                {/* forceRender to allow the field values to flow to form submit */}
+                                <Panel key="1"  header="Advanced settings" forceRender>  
+                                    <LockTimeSelector />
                                 </Panel>
-                                </Collapse>
-                            <Row>
+                            </Collapse>
+                            <Row> 
                                 <Col span={24}>
                                     <Button
                                         loading={isPending}
