+++ src/5_shared/ui/LockTimeSelector/index.ts
@@ -1,0 +1,1 @@
+export { LockTimeSelector } from './LockTimeSelector'
